net: lib: azure_fota: Add option to auto validate modem firmware
Add option to auto validate modem firmware at sys init.
