+++ include/net/azure_fota.h
@@ -80,6 +80,14 @@ int azure_fota_init(azure_fota_callback_t evt_handler);
  */
 int azure_fota_msg_process(const char *const buf, size_t size);
 
+/** @brief Validate modem firwmare update.
+ *	   This function must be called before any other modem related functionality is called.
+ *	   If required, the device will be rebooted to complete the modem update.
+ *	   By default, CONFIG_AZURE_FOTA_MODEM_FW_UPDATE_AUTO_VALIDATE is set which will cause
+ *	   modem initialization handling to occur automatically, before the application is run.
+ */
+void azure_fota_modem_firmware_update_validate(void);
+
 #ifdef __cplusplus
 }
 #endif
